+++ src/system/Mount.hxx
@@ -8,4 +8,4 @@
  * Throws std::system_error on error.
  */
 void
-BindMount(const char *source, const char *target, int flags);
+BindMount(const char *source, const char *target, unsigned long flags);
